backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
Django = "^4.1.3"

[tool.poetry.dev-dependencies]
tryceratops = "^1.1.0"
isort = "^5.10.1"
flake8 = "^6.0.0"
black = "^22.10.0"
mypy = "^0.991"
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-watch = "^4.2.0"
factory-boy = "^3.2.1"
Faker = "^15.3.3"
pytest-factoryboy = "^2.5.0"
django-admin = "^2.0.1"
djhtml = "^1.5.2"
pytest-django = "^4.5.2"
requests-mock = "^1.10.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
extra_standard_library = ["pytest", "toml", "click"]

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
